[AIPLATFORM-2154] chore: 변수명 변경, 메시지 const변경, send_mail_v2 -> send_mail 파일명 변경
